Modify/Delete/Images

+++ src/components/CustomHooks/useDelete.js
@@ -1,23 +1,43 @@
 import axios from "axios"
 import { delete_category, delete_product } from "../../redux/actions/sync"
+import swal from "sweetalert"; 
 
 const baseUrl = "http://localhost:3001/api/v1"
+
+
 
 export default function useDelete (dispatch) {
 
     const handleDelete = (type, id, imgPath) => {
+        let name = "Category"
+        
         if(type === "products") {
-            axios.delete(" http://localhost:3001/api/v1/products/image/delete", {
-                completeRoute: imgPath
-            }).then(res=> res.json()).then(res=> console.log(res)).catch(err=> console.log(err))
+            name = "Product"
+        }
+            swal({
+                title: "Are you sure?",
+                text: `Once deleted, you will not be able to recover the ${name}!`,
+                icon: "warning",
+                buttons: true,
+                dangerMode: true,
+              })
+              .then((willDelete) => {
+                if(willDelete) {
+                axios.delete(`${baseUrl}/${type}/${id}`);
+                return true
+            } else return false
+              }
+              ).then((res)=> {
+                if(res)  {
+                    swal(`The ${name} is deleted!`, {
+                    icon: "success"})
+                    type=== "categories" ? 
+                    dispatch(delete_category(id))
+                    :dispatch(delete_product(id))}
+                else { swal(`Your ${name} is safe!`); }
 
-            dispatch(delete_product(id))
-       } else if(type === "categories") {
-           dispatch(delete_category(id))
-       } 
-        axios.delete(`${baseUrl}/${type}/${id}`)
-        .then(()=> console.log("borrado XD"))
-        .catch((err)=> console.log(err))
+            }
+              ).catch(err=> console.log(err))
         
     }
